+++ tests/integration_tests.rs
@@ -6,9 +6,10 @@ use std::path::PathBuf;
 use tempfile::tempdir;
 use test_shared::*;
 
-fn test_scanner_config() -> ScannerConfig {
+fn test_scanner_config(with_git: bool) -> ScannerConfig {
     let mut config = ScannerConfig::default("test");
     config.data_id = Some("test-id".to_string());
+    config.features.git = with_git;
     config
 }
 
@@ -19,7 +20,8 @@ fn it_calculates_lines_of_code() -> Result<(), Error> {
     let mut buffer: Vec<u8> = Vec::new();
     let out = Cursor::new(&mut buffer);
 
-    let result = polyglot_code_scanner::run(&root, &test_scanner_config(), None, &["loc"], out);
+    let result =
+        polyglot_code_scanner::run(&root, &test_scanner_config(false), None, &["loc"], out);
 
     assert!(result.is_ok());
 
@@ -41,7 +43,8 @@ fn it_calculates_git_stats() -> Result<(), Error> {
     let mut buffer: Vec<u8> = Vec::new();
     let out = Cursor::new(&mut buffer);
 
-    let result = polyglot_code_scanner::run(&git_root, &test_scanner_config(), None, &["git"], out);
+    let result =
+        polyglot_code_scanner::run(&git_root, &test_scanner_config(true), None, &["git"], out);
 
     assert!(result.is_ok());
 
@@ -63,8 +66,8 @@ fn it_calculates_detailed_git_stats() -> Result<(), Error> {
     let mut buffer: Vec<u8> = Vec::new();
     let out = Cursor::new(&mut buffer);
 
-    let mut config = test_scanner_config();
-    config.detailed = true;
+    let mut config = test_scanner_config(true);
+    config.features.git_details = true;
 
     let result = polyglot_code_scanner::run(&git_root, &config, None, &["git"], out);
 
@@ -80,32 +83,4 @@ fn it_calculates_detailed_git_stats() -> Result<(), Error> {
     Ok(())
 }
 
-// TODO: THIS TEST IS JUNK! We need some commits that actually generate coupling data :)
-
-#[test]
-fn it_calculates_detailed_git_stats_with_coupling() -> Result<(), Error> {
-    let gitdir = tempdir()?;
-    let git_root = unzip_git_sample("git_sample", gitdir.path())?;
-
-    let mut buffer: Vec<u8> = Vec::new();
-    let out = Cursor::new(&mut buffer);
-
-    let mut config = test_scanner_config();
-    config.detailed = true;
-    let coupling_config =
-        polyglot_code_scanner::coupling::CouplingConfig::new(3, 1, 0.1, 120 * 60, 60 * 60, 0, None);
-
-    let result =
-        polyglot_code_scanner::run(&git_root, &config, Some(coupling_config), &["git"], out);
-
-    assert!(result.is_ok());
-
-    let parsed_result: Value = serde_json::from_reader(buffer.as_slice())?;
-
-    assert_eq_json_file(
-        &parsed_result,
-        "./tests/expected/integration_tests/git_detailed_flare_test.json",
-    );
-
-    Ok(())
-}
+// TODO: add a coupling e2e test!  Needs a lot of setup